refactor: extract whois patterns data
